+++ src/cone/maps/browser/static/maps/cone.maps.js
@@ -1,16 +1,6 @@
-(function (exports, $) {
+var cone_maps = (function (exports, $) {
     'use strict';
 
-    function lookup_factory(name) {
-        let ob = window;
-        for (let part of name.split('.')) {
-            ob = ob[part];
-            if (ob === undefined) {
-                throw "Cannot locate map factory: " + name;
-            }
-        }
-        return ob;
-    }
     let layer_factories = {};
     layer_factories.tile_layer = function(inst, cfg) {
         inst.layer_created(new L.TileLayer(cfg.urlTemplate, cfg.options), cfg);
@@ -23,23 +13,26 @@
     class Map {
         static initialize(context) {
             $('div.cone-map', context).each(function() {
-                let elem = $(this);
-                let factory = lookup_factory(elem.data('map-factory'));
-                new factory(elem);
+                let elem = $(this),
+                    settings = elem.data('map-settings'),
+                    factory_path = settings.factory,
+                    factory = ts.object_by_path(factory_path);
+                new factory(elem, settings);
             });
         }
-        constructor(elem) {
+        constructor(elem, settings) {
             this.elem = elem;
             this.id = elem.attr('id');
-            this.layers = elem.data('map-layers');
-            this.default_center = elem.data('map-center');
-            this.default_zoom = elem.data('map-zoom');
-            this.map_options = elem.data('map-options');
-            this.control_options = elem.data('map-control-options');
-            this.markers = elem.data('data-map-markers');
-            this.markers_source = elem.data('data-map-markers-source');
-            this.marker_groups = elem.data('data-map-groups');
-            this.marker_groups_source = elem.data('data-map-groups-source');
+            this.layers = settings.layers;
+            this.default_center = settings.center;
+            this.default_zoom = settings.zoom;
+            this.default_bounds = settings.bounds;
+            this.map_options = settings.options;
+            this.control_options = settings.control_options;
+            this.markers = settings.markers;
+            this.markers_source = settings.markers_source;
+            this.marker_groups = settings.groups;
+            this.marker_groups_source = settings.groups_source;
             this.create();
             elem.data('map-instance', this);
         }
@@ -47,15 +40,20 @@
             this.create_map();
             this.create_controls();
             this.create_layers();
+            this.create_markers();
         }
         create_map() {
             this.map = new L.Map(this.id, this.map_options);
-            this.map.setView(this.default_center, this.default_zoom);
+            if (this.default_bounds.length) {
+                this.map.fitBounds(this.default_bounds);
+            } else {
+                this.map.setView(this.default_center, this.default_zoom);
+            }
         }
         create_controls() {
             let base_maps = [],
                 overlay_maps = [];
-            this.map_layers = new L.control.Layers(
+            this.map_layers = new L.Control.Layers(
                 base_maps,
                 overlay_maps,
                 this.control_options
@@ -84,6 +82,30 @@
         remove_layer(layer) {
             this.map.removeLayer(layer);
         }
+        create_markers() {
+            for (let marker of this.markers) {
+                this.create_marker(marker);
+            }
+            if (this.markers_source) {
+                $.getJSON(this.markers_source, function(data) {
+                    for (let marker of data) {
+                        this.create_marker(marker);
+                    }
+                }.bind(this));
+            }
+            if (this.markers || this.markers_source) {
+                this.map.on('popupopen', function(evt) {
+                    let popup = evt.popup;
+                    ts.ajax.bind($(popup._contentNode));
+                });
+            }
+        }
+        create_marker(marker) {
+            let m = new L.Marker(marker.latlng, marker.options).addTo(this.map);
+            if (marker.popup) {
+                m.bindPopup(marker.popup.content, marker.popup.options);
+            }
+        }
     }
 
     $(function() {
@@ -96,13 +118,8 @@
 
     exports.Map = Map;
     exports.layer_factories = layer_factories;
-    exports.lookup_factory = lookup_factory;
 
     Object.defineProperty(exports, '__esModule', { value: true });
-
-
-    window.cone_maps = exports;
-
 
     return exports;