+++ CHANGES.rst
@@ -4,7 +4,29 @@ Changes
 0.2 (unreleased)
 ----------------
 
-- No changes yet.
+- Add ``Path.Drag.js`` to resources.
+  [rnix]
+
+- Add ``Leaflet.Editable`` to resources.
+  [rnix]
+
+- Map settings are defined via ``MapTile.map_settings`` property.
+  [rnix]
+
+- Map settings gets rendered as single data attribute.
+  [rnix]
+
+- Add ``MapTile.map_bounds`` property.
+  [rnix]
+
+- Fix control layers creation in ``Map.create_controls``.
+  [rnix]
+
+- Implement map markers from JSON source.
+  [rnix]
+
+- Implement static map markers.
+  [rnix]
 
 
 0.1 (2021-11-21)